prima: Fix for issue RIVA does not seem to be getting into BMPS mode.
The reason riva does not enter in to power save is because after the
connection to the AP (Configured 40Mhz channel in 2.4Ghz) the driver
sends the invalid cb mode in the finish scan request to the Riva if we
are in scan window. The issue does not show up if we are not in scan
window.

The reason driver sends the invalid cb mode is because after receiving
the join beacon the variable psessionEntry->htSecondaryChannelOffset
in the function limCheckAndAnnounceJoinSuccess is over writing with
the AP secondary offset.