prima: Release 3.2.1.1
The following changes were merged in this release:

648542: Scan optimizations to support concurrency
New Features:
1. Sessionize SME scan cache
2. Separate beacon & porbe rsp entry for PE scan cache
3. Enable scan break down logic in SME for multi-session
Fixes:
The change fixes issue with host initiated scans interferring with WFD
connection and causing jitter in the video stream on the WFD
connection

648544: IMPS mode, TX COMP interrupt flooding fix
To stop TX COMP interrupt with IMPS mode, RIVA shuld power down. But
whatever reason, if RIVA could not power down, TX COMP INT could not
be cleared and infinite INT may happen

648399: Fix for issue RIVA does not seem to be getting into BMPS mode
The reason riva does not enter in to power save is because after the
connection to the AP (Configured 40Mhz channel in 2.4Ghz) the driver
sends the invalid cb mode in the finish scan request to the Riva if we
are in scan window. The issue does not show up if we are not in scan
window.
The reason driver sends the invalid cb mode is because after receiving
the join beacon the variable psessionEntry->htSecondaryChannelOffset
in the function limCheckAndAnnounceJoinSuccess is over writing with
the AP secondary offset.