prima: CSR pre-auth timer is made larger than Lim timer
CSR and Lim timers are of same time, 1 sec. because of which, CSR
timer starts first and Lim timer next, if the pre-auth response is not
received, then CSR timer expires first and sends next pre-auth
request, for which LIM timer is still running,it does not start a new
timer and it expires.

Which sends a pre-auth status with error and later for the 2nd
pre-auth request, STA receives pre-auth response also.  First error
response is cleaning up the context and it crashes for the next
successful pre-auth