prima: Fix lock used while doing any modification on vos packet lists.
In the current implementation we are using mutex_lock_interruptible
while trying to access the vos packet lists now made the change to
make use of mutex_lock insted of mutex_lock_interruptible so that the
task doesn't get interrupted while trying to modify the list.

Also in while releasing the packet back to free list we were just
ignoring whether we have succesfully acquired the lock or no and we
are trying to access the list which could result in the corruption of
the list.so made sure we access the list after acquiring the lock.