Subsystem restart invoke mutex_lock in atomic environment
When WCTS_NotifyCallback() is invoked in the interrupt context,to
process 'SMD_EVENT_CLOSE' event it is trying to get the VOS List mutex
to clear the transport message queue which is not desirable in the
interrupt context.  To fix the issue moved the cleaning of the
transport message queue to wdi_shutdown().