Changes to acquire wakelock to avoid the apps suspend during the scan.
Acquire the wakelock to avoid the apps suspend during the scan. To
address the following issues.
1) Disconnected scenario: we are not allowing the suspend as WLAN is
not in BMPS/IMPS this result in android trying to suspend aggressively
and backing off for long time, this result in apps running at full
power for long time.
2) Connected scenario: If we allow the suspend during the scan, RIVA
will be stuck in full power because of resume BMPS.