Reduce the age count only for valid scan entries
Scan aging was reducing the age count of all scan records, even ones
which had been aged out.  This resulted in a "modify after free" which
was detected by kernel SLUB debugging.  Modify the code to reduce the
age count only if the scan entry is not getting aged out already.

b/7249823