wlan: Release 3.2.1.5b
The following change was incorporated in this release:

---
Reduce the age count only for valid scan entries

Scan aging was reducing the age count of all scan records, even ones
which had been aged out.  This resulted in a "modify after free" which
was detected by kernel SLUB debugging.  Modify the code to reduce the
age count only if the scan entry is not getting aged out already.

b/7249823