Preferentially scan occupied channels (learned from previous scans).
Driver initiated "background scans" used to use a scan channel list
from "cfg.ini".  Now, the list is formed from the valid channel
list. And a record is kept of the subset of the valid channels that
are occupied.  Subsequent background scans will preferentially scan
the occupied channels, followed by the unoccupied channels.