Enable gEnableBypass11d to reduce the connection time.
The sta takes 6 to 7 secs to connect to an AP when we turn off and
turn on Wi-Fi the reason is during the driver load time.  If the
gEnableBypass11d flag is seto zero then the driver does the passive
scan for all the channels. So to reduce the connection time set the
gEnableBypass11d flag to 1 in the ini file.