wlan: Release 3.2.1.6
The following changes were incorporated in this release:
---
Fix for passing proper reason code to supplicant

Passing reason code to supplicant based on reason code which received
during deauth frame.
---
Update template cfg.ini for LFR

Update the gNeighborLookupThreshold to 70 and
gNeighborReassocThreshold to 75.
---
Preferentially scan occupied channels (learned from previous scans).

Driver initiated "background scans" used to use a scan channel list
from "cfg.ini".  Now, the list is formed from the valid channel
list. And a record is kept of the subset of the valid channels that
are occupied.  Subsequent background scans will preferentially scan
the occupied channels, followed by the unoccupied channels.
---
Remove gNeighborScanChannelList from cfg.ini

The driver-initiated background scans now use the valid channel list
instead of gNeighborScanChannelList from cfg.ini to determine the scan
channel list.  Therefore, this item is removed from cfg.ini.
---
Enable gEnableBypass11d to reduce the connection time.

The sta takes 6 to 7 secs to connect to an AP when we turn off and
turn on Wi-Fi the reason is during the driver load time.  If the
gEnableBypass11d flag is seto zero then the driver does the passive
scan for all the channels. So to reduce the connection time set the
gEnableBypass11d flag to 1 in the ini file.
---
Inform TL about pending packets whenever Queue is filling up

Currently, HDD informs TL about pending packets only when it goes
empty to non-empty.  If TL misses this signal, it will not hear back
again from HDD, and the queue will fill and never be serviced.  So, to
avoid this, we inform TL whenever the queue is getting filled
---
Update the Channel Change Reason

Update the channel change reason to the default generic channel switch
reason before initiating channel switch.  If this is not done, the
channel change reason is "0" which is LIM_SWITCH_CHANNEL_REASSOC and
is leading to a reassociation.
---
Set RoamRssiDiff=0 to disable roam hysteresis.

Roam hysteresis setting is to be included in a future build.  For now,
Set RoamRssiDiff=0 to disable it.
---
Stop scan refresh timer while resetting report scan state

1) Refresh scan results timer fires in wrong state: When pre-auth
fails three times and there are no other roam candidates, we start
ResultsRefreshTimer which is 20s. In the meantime, we could receive a
deauth from the current AP. But, we do not stop the
ResultsRefreshTimer in csrNeighborRoamResetReportScanStateControlInfo.
Due to this, the timer fires when CSR state machine is in INIT state
and tries to scan by transitioning to CFG_CHAN_SCAN. This by itself is
a wrong state transition. In addition, we clear session ID related
info in the timers when we received the deauth. Due to this, APQ
freezes (the screen goes blank and no access to adb shell either) and
happens every time the ResultsRefreshTimer fires after 3 pre-auth
failures. This is fixed by stopping the timer in
csrNeighborRoamResetReportScanStateControlInfo.

2) Outdated results in scan list prevents roam candidate selection and
leading to heartbead failure: At a site where there are more than 150
APs, roam candidate selection fails because the scan list is filled
with outdated results. This is fixed by: a) increasing the size of
scan list to 200, and b) purging old results every time we begin a
fresh scan cycle (by definition, a scan cycle is hopping through
multiple channels in search of roam candidates). Our tests showed that
(b) alone was sufficient, but (a) would help in instances where we
could find more than 100 scan results within each scan cycle.
---
Fix for setting properly the AC mask for TxConn

In certain circumstances TL would misscalculate the AC mask and
pending flag and wrongfully stop servicing the EAPOL frame queue from
HDD.  Fix the code to correct set the AC mask and pending flag.