Increase number of buckets to categorize roamable APs
Roamable candidates found after scanning are categorized into buckets
based on their respective RSSI. Currently, width of each bucket is
defaulted to 30 and the number of buckets is set to 5. Due to this,
all roam candidates within -60 to -90 dBm fall into a single
bucket. Roam candidates from the same bucket are then sorted based on
their capabilities. Hence, it is possible that we might pick an AP
with -89 dBm ahead of another candidate AP with -61 dBm.  This problem
is fixed by increasing the number of buckets and reducing the width
between the buckets. This check-in brings in support for a width of 5
dBm (configurable through ini parameter gRssiCatGap) and increases the
number of buckets to 15.