wlan: Release 3.2.1.8d
The following changes were incorporated in this release:
===
Prevent P2P from flushing scan results while roaming

If we are in the process of roaming, don't allow P2P to flush the
current scan results.  Otherwise the roaming will fail.
===
Even BMPS mode, display correct RSSI within TL log MSG

Previously TL RSSI logging was displaying incorrect values.  Update
the code so that the correct RSSI values are displayed
===
Increase number of buckets to categorize roamable APs

Roamable candidates found after scanning are categorized into buckets
based on their respective RSSI. Currently, width of each bucket is
defaulted to 30 and the number of buckets is set to 5. Due to this,
all roam candidates within -60 to -90 dBm fall into a single
bucket. Roam candidates from the same bucket are then sorted based on
their capabilities. Hence, it is possible that we might pick an AP
with -89 dBm ahead of another candidate AP with -61 dBm.  This problem
is fixed by increasing the number of buckets and reducing the width
between the buckets. This check-in brings in support for a width of 5
dBm (configurable through ini parameter gRssiCatGap) and increases the
number of buckets to 15.
===
No Scan Interference During HandOff

When a Handoff is in progress, during preauth or report scan or
reassoc states, a scan request will be denied with a busy signal.
===
Flush only P2P results during p2p_find
===