Ensure pre-auth is processed before invoking HDD callbacks
This fix addresses a race condition between roaming engine trying to
pre-auth and PE/LIM trying to disconnect/tear down the link due to
heartbeat failure. In this scenario, we clear our state machine upon
receiving the disconnect indication. Then, the pre-auth to the new AP
is sent by the PE to which we get a successful response. When the
pre-auth response notification reaches SME (in particular, CSR
neighbor roaming code), we are continuing to process and triggering a
roam even though we have disconnected and are in INIT state.  This is
addressed by having csrNeighborRoamPreauthRspHandler routine return a
status whether the pre-auth has been processed. If it is not even
processed (because we received the pre-auth in a wrong state), then we
need to silently drop it.