wlan: Release 3.2.1.9b
The following changes were incorporated in this release:
===
Ensure pre-auth is processed before invoking HDD callbacks

This fix addresses a race condition between roaming engine trying to
pre-auth and PE/LIM trying to disconnect/tear down the link due to
heartbeat failure. In this scenario, we clear our state machine upon
receiving the disconnect indication. Then, the pre-auth to the new AP
is sent by the PE to which we get a successful response. When the
pre-auth response notification reaches SME (in particular, CSR
neighbor roaming code), we are continuing to process and triggering a
roam even though we have disconnected and are in INIT state.  This is
addressed by having csrNeighborRoamPreauthRspHandler routine return a
status whether the pre-auth has been processed. If it is not even
processed (because we received the pre-auth in a wrong state), then we
need to silently drop it.
===
Handle disconnects in reassoc state

This fix addresses a bug wherein we could be stuck in reassoc state
forever, thereby preventing all scans.

A previous change added logic to block all HDD-triggered scans during
one of the four handoff states: report scan, pre-auth, pre-auth done
and reassoc. This was done to protect roaming algorithm from getting
into erroneous states during concurrency operation wherein P2P "remain
on channel" mode can inhibit seamless roaming.

Our current code could get stuck in reassoc state if we disconnect
while we are in that state. This is because we do not transition to
INIT state in csrNeighborRoamIndicateDisconnect from reassoc
state. However, it is OK to transit provided we are NOT in Disassoc
Handoff sub-state.  This is because we receive disconnect notification
for the old AP in Disassoc Handoff sub-state. So, any disconnect
indications received in sub-states other than Disassoc Handoff should
be treated as a trigger to transition from reassoc state to INIT.
===