Unable to connect if the AP sends country code as "I".
The problem was if the AP sends the country code as "I" then the
function csrMatchCountryCode compares the list of country code with
the country code present in the AP beacon. Since this country code is
not present the csr fails to connect to that AP which is causing the
problem.  The csrMatchCountry code has an issue that there is no need
for validation unless the fEnforceDefaultDomain and
fEnforceCountryCodeMatch is set to true in the WCSS_qcom_cfg.ini.