Dampen scan periodicity if no roamable candidates found.
This fix limts the iterations of background scans to find potential
roam candidates in scenarios where there are none (home environment or
enterprise network with coverage holes).

When no roamable candidates are found, we register with a new lookup
threshold and continue to scan if the RSSI of the current AP falls
below the new threshold. However, if we are in an area where the
current AP is the only available candidate and RSSI is below the
threshold, we would endlessly continue to scan. In the interest of
power consumption, we would like to dampen the scans every time there
are no roamable candidates found.

For every scan that results in no candidates, double the scan
periodicity (initialized to 1s) until we hit a max (defined as
60s). Subsequently, scan every 60s if we continue to not find any
candidates. Once a candidate is found, the periodicity is reset back
to 1s.