wlan: Release 3.2.1.10
The following changes were incorporated in this release:
===
Unable to connect if the AP sends country code as "I".

The problem was if the AP sends the country code as "I" then the
function csrMatchCountryCode compares the list of country code with
the country code present in the AP beacon. Since this country code is
not present the csr fails to connect to that AP which is causing the
problem.  The csrMatchCountry code has an issue that there is no need
for validation unless the fEnforceDefaultDomain and
fEnforceCountryCodeMatch is set to true in the WCSS_qcom_cfg.ini.
===
Lowered the log level of a WDI print

Fixes misleading print in error level
===
Dampen scan periodicity if no roamable candidates found.

This fix limts the iterations of background scans to find potential
roam candidates in scenarios where there are none (home environment or
enterprise network with coverage holes).

When no roamable candidates are found, we register with a new lookup
threshold and continue to scan if the RSSI of the current AP falls
below the new threshold. However, if we are in an area where the
current AP is the only available candidate and RSSI is below the
threshold, we would endlessly continue to scan. In the interest of
power consumption, we would like to dampen the scans every time there
are no roamable candidates found.

For every scan that results in no candidates, double the scan
periodicity (initialized to 1s) until we hit a max (defined as
60s). Subsequently, scan every 60s if we continue to not find any
candidates. Once a candidate is found, the periodicity is reset back
to 1s.
===