Add dummy handlers for Android PNO commands
Currently when the WLAN driver is presented with Android standard
PNOSETUP and PNOFORCE commands, the driver processses them as if they
are our proprietary PNO command.  Since the data provided by the
standard commands doesn't match the format of the proprietary command,
unexpected driver behavior may result.  Prevent this by adding dummy
handlers for the PNOSETUP and PNOFORCE commands so that those specific
commands don't result in invoking the proprietary PNO command handler.