Use unique logging for driver commands
There are two different mechanisms by which driver commands may be
sent, either by Wireless Extensions private ioctls or by regular
device private ioctls.  Currently when processing a driver command the
log message does not indicate via which mechanism the command was
received.  Since knowing this information aids in debugging, modify
the log messages so that they indicated the function that is handling
the message.