wlan: Release 3.2.1.11a
The following changes were incorporated in this release:
===
Handle race condition in Management Frame TX Timeout

When WDA times out waiting for the TX path to send out a management
frame it informs PE about the failure ; PE destroys the frame in the
MC thread context In the mean while the frame can be anywhere in the
TX data path components and its deletion will cause an unexpected
behavior that will more often then not lead to a crash.

The proposed fix prevents WDA from notifying failure to PE, instead it
tags the frame as timed out using the WDA user data id from the vos
packet.

When frame completes tx it will end up in WDA marked appropriately -
WDA will just discard the packet at its level and it will not do
anymore processing as this isn't the current frame that it is trying
to send out.

Unblocking the MC thread before a frame has completed tx opens up the
possibility of UMAC sending another mgmt frame to TL while the 1st one
is still pending TX at TL level. In this case to eliminate any
possible race condition and not to introduce locks TL will just fail
the TX request.

In all customer reports the crash seems to happen when the frame was
already on its completion path and the time gap between the timeout
and crash is very small, reason for which I also increased the WDA
timer value to minimize the likelyhood of this occurance.
===
Fix for crash during roaming

As per Stability team's analysis, changing the type of variable "len"
from unsigned to signed in limGetBssDescription() which may be the
reason for crash during roaming.  We do a memcopy based on this
variable which is updated gradually within the function to extract
different information from the buffer provided.
===