Ignore roam candidates with RSSI lower than lookup threshold
Roaming to a candidate with low RSSI is triggering a series of
undesirable events which are consuming lot more power. If we
roam to a candidate with RSSI lower than lookup threshold, we are
unable to enter BMPS (possibly due to difficulty exchanging null
data). Due to this we remain active and immediately get notified
by TL since RSSI is lower than lookup threshold to begin with. This
shall trigger background scans to find new candidates.

This problem is being fixed by ignoring candidates whose RSSI
is lower than lookup threshold.