Teardown Link with AP if riva indicates del sta reason keep alive
There are two problems associated with this change
1) When Riva sends the delete station context indication to the Host
with reason code HAL_DEL_STA_REASON_CODE_KEEP_ALIVE and the assocID
other than the value 1 then PE does not indicate the disconnect
indication to SME.  The fix is in case of STA mode, RA deals with peer
STA which happens to be a BSS (STA Type = Peer), so we should be
ignoring the assoc ID as it has no meaning.  It has meaning only for
STA Type = SELF.
2) BTC module dynamically enable/disable HB monitoring for specific BT
events but PE does not reactivate the Heart beat timer when the BTC
module enable the Heart beat.The fix is to reactivate the Heart beat
timer.