wlan: Release 3.2.1.11i
The following change was incorporated in this release:
===
Post proper response to SME when reassoc is rejected

When AP rejects reassoc request (i.e. status code other than 0), we
are posting an incorrect response (i.e. eWNI_SME_JOIN_RSP) to SME in
limSendDelStaCnf. Upon receiving this message, SME silently drops it
because it is in eCSR_ROAM_SUBSTATE_REASSOC_REQ sub-state and
expecting only eWNI_SME_REASSOC_RSP. Due to this, the upper layers do
not get informed about the link being torn down and the roaming engine
is also stuck in reassociating state. This leads to tx_timeouts being
reported by HDD.

This is being fixed by including eSIR_SME_FT_REASSOC_FAILURE in
limSendDelStaCnf so that correct response is posted to SME.