wlan: Release 3.2.1.12
The following changes were incorporated in this release:
===
Update restart driver API to support statically linked WLAN driver

When WLAN driver is statically linked, the restart driver API will
invoke the WCNSS reset interrupt API to initiate SSR.
===
Notify the Kernel suspend to WCNSS

Notify WCNSS when WLAN driver has suspended; so that WCNSS need not
send certain notifications to Apps. It is only required to do this
when driver is in BMPS mode.
===
Conditionally invoke driver restart API

The API to restart the WLAN driver may not be available in all
kernels, so only try to invoke it if the kernel supports it.
===