De-register lookup UP indication during pre-auth
Upon receiving a lookup UP indication, it suggests that the link to the
current AP has improved and we clean-up all scan control info before
moving back to connected state. However, if we receive a lookup UP
indication while we are in the middle of a pre-auth, it can lead to
several race conditions:

1) Lookup UP indication received after SME posted pre-auth message
to PE and before PE posts the auth response back to SME.
In this instance, while SME layer cancels the pre-auth, the PE/LIM
would have created an FT session which goes unaccounted for. This extra
session causes subsequent scans to fail and eventually results in a
HB failure.

2) Lookup UP indication received after reassoc and before key exchange.
This instance is even more severe. When this happens, we would ignore
the connect indication received following key exchange. Due to this,
all layers except the neighbor roam engine has moved to the new AP
while the roam engine still thinks we are associated to the older AP.
Due to this mis-matched roam profile, we end up: (i) rejecting new
candidate APs found during scan, (ii) pre-auth with the AP we are
currently associated to, etc.

This is being fixed by de-registering for lookup UP indication
as soon as we receive the reassoc DOWN indication. If the pre-auth
fails with all available roam candidates, we once again register
for a lookup UP indication.