wlan: Stop hdd_restart_timer at unload only if it is not FTM mode
VOS_ASSERT is observed at driver unload in FTM mode. hdd_restart_timer
is only initialized at wlan_startup when the mode is not FTM but is
stopped at unload unconditionally thus casuing an assert for
stopping an uninitialized timer. The fix is to stop the timer in
wlan_exit only if the mode is not FTM

Change-Id: Ie06fd3ebc1710b14e441474de8553f64b66b6797
CR-Fixed: 433127