wlan: remove obsolete RTL8652 and LX5280 conditional code
The current driver source contains obsolete conditional code, some
that was originally used to optimize the performance of the AP code on
the RTL8652 platform, and some that was used to customize the code on
the LX5280 projector.  Since those platforms are no longer supported,
remove the conditional code for them.

Change-Id: I0ba4726a03825a47fe5fc917ccc7081f66af775d
CRs-fixed: 452041