wlan: hdd: remove obsolete "INTEGRATED_SOC" featurization
The current driver is designed around an "integrated SoC" solution
where the WLAN processor and the Application Processor are integrated.
But the driver is based upon a prior driver that supported a
non-integrated discrete solution.  During the time the driver was
converted from a non-integrate to integrated solution conditional
compilation was introduced so that both architectures could be
supported by the same driver.  But the non-integrated solution is no
longer supported by this driver, so remove all of the conditional
compilation so that the current driver will be easier to support.

Change-Id: Ic8ba8fb1d57ad501eac6be34b1a6b8c4e33507b3
CRs-fixed: 452041