wlan: validate key length passed to sme_RoamSetKey()
Static source code analysis identified an issue where the key length
being passed to sme_RoamSetKey() was not being validated.  As a result
the operations used to display and copy the key could result in a
buffer overflow.  To prevent buffer overflow add logic to validate the
key length before using it to display or copy the key.

Change-Id: I4e5fadd2d65e069f8a61ef5bf99b157ffa0ee98b