wlan: Fix memory leak in csrInitCountryValidChannelList()
Static source code analysis identified a memory leak in
csrInitCountryValidChannelList().  This function allocates memory for
pOutChannelList but it never saves the pointer, and hence the memory
is leaked (and the function doesn't actually perform the function it
was designed to perform).  Update the code so that the pointer to the
channel list is actually saved for later use.

Change-Id: Iac148ae15458332279705c2da738d5f43635724c
CRs-fixed: 470188
(cherry picked from commit 45ebd2ab5b9886962937d0736029da428d025017)