wlan: Zero Value Initialization of Memory.
Correction of Zero value Initialization.
If the memory is not initialized to Zero
values, there is a high chance of sending
down inappropriate values to the firmware.
The FW may consume the values and it could
lead to a stability issue.

Change-Id: I27ef9400f70dbc730e0845c57ef356a342322706
CRs-Fixed: 493853
(cherry picked from commit 7cfcd6119234e39dff795c0c5a843df65272b117)