wlan: Do not wait for supplicant to transit the STA to authorized.
wlan: fix for WEP with EAP-SIM/AKA method insists to wait
for the supplicant to set the STA to authorized state.
Relying on Supplicant to move the station to authorized results
in the delay in 11R/LFR roam times. This commit shall not
depend on supplicant to move the station/port to the authorized
mode.

CRs-fixed: 490485
(cherry picked from commit 24a339b225b2f6cb7dab8bd3e5de6f6a22000173)

Change-Id: I4cc0f41294f9d3f0578338c5428295e7233bb037