wlan: Add traffic monitor on SAP/P2P-GO mode
To address the issue wherein APPS fails to power collapse
when OBSS protection is enabled,
we need to implement an IDLE traffic monitoring
logic in the HDD: it needs to detect 2 states:
IDLE STATE: HDD transitions to that STATE
when there is no traffic for a configuration duration in cfg file.
ACTIVE STATE: this is the initial state when SAP comes up.
HDD moves also to that state
when it sees traffic when in IDLE State.

CRs-fixed: 495899
(cherry picked from commit e45dd304cb46665223cce1d6dc118f77766fd498)

Change-Id: Idbd672c8a1bb39630aff0c81f84937f4668715ba