wlan: no need to re-enable imps/bmps during driver unload
In the disassoc handler we are re-enabling imps & bmps if we had
multisession. In driver unload case, this is not needed. In fact it
creates a situation where hal stop is sent to FW while it is already
in IMPS. Ths fix is to avoid that.

Change-Id: Ib8c2058a71b048a7b4d029f8b533a7cba4334e86
CRs-fixed: 496903
(cherry picked from commit 80c84dc18eec41ce3bb299723dd4841f8c7d1a84)