wlan: Change RX wakelock to 50 milliseconds
When the original change to add wakelock to the RX path was added, we
intended to use only 50 milliseconds. However msecs_to_jiffies() were
not added to the timeout argument; so later when the API was corrected
it was considered as 500 milliseconds. Restore the original intended
value.

CRs-Fixed: 498941
(cherry picked from commit 8545cfa54e085cbf70988933e06febfbf799ff17)

Change-Id: I37213f26c3b862159acf296826a48c7cad817023