wlan: Do not depend on supplicant to transit the STA to authorized
wlan: Commit "Do not wait for supplicant to transit the STA to
authorized" insists not to depend on the supplicant to do so.
Thus removing the functionality in wlan_hdd_change_station to
achieve the same.

Change-Id: Ic418952d2a6435a3d77648a47b03362dccdf136c
CRs-fixed: 498232
(cherry picked from commit 09ea78840481a719bf30f20d1052e0466b03e9b3)