wlan: remove duplicate IF condition variable checks
The gerrit changes 380571 are done for the jb_mr2 branch
which does not have all the fixes from master branch.
With the gerrit 380571 on master branch introduces duplicate IF
condition which are already handled in hdd_set_pwrparams.
The fix is to remove checks which are not required and
already present in hdd_set_pwrparams

Change-Id: Ia75955893ded5963757712adab3ef6e63bea5a38
CRs-Fixed: 513043