wlan: Prevent HDD roam profile being cleared during assoc process
During the normal course of association, SME returns the "roamStatus" of
eCRS_ROAM_SHOULD_ROAM.  Currently HDD treats this as an opportunity to
clear its copy of the "roam profile".  While this is fine for all other
cases, this doesn't always work for WEP associations.  Because for WEP,
since the "network profile" supplied by the supplicant isn't specific
about "WEP open" or "WEP shared" HDD has to try both.  But since HDD
copy of the "roam profile" is cleared during the initial association,
this means that the re-try of the WEP assoc will be to a "cleared" or
"wildcard" SSID.

Change-Id: I3594e75e85e5a8b1aa6c2390c25f1f2d9d8b7929
CRs-fixed: 557427