wlan: fixed a memory leak when driver is unloading.
Remove the csrStop routine call in csrClose function as csrStop
is already called in smeStop routine during the driver unload.
Calling csrStop in csrClose routine leads to memory leak during
the driver unload as csrStop posts message to PE and the message
will not be processed since the scheduler is already closed in
the unload process.

Change-Id: Id2097875c0e3b9b9ceeb8d3c5c7ea8c4db0a4ea4
CRs-fixed: 579445